Move constances out of types
