+++ public/app/features/provisioning/GettingStarted/GettingStarted.tsx
@@ -1,190 +1,54 @@
-import { useNavigate } from 'react-router-dom-v5-compat';
-
-import { LinkButton, Alert, Stack, Text, Button, Box, Icon, IconName } from '@grafana/ui';
+import { Alert, Stack, Text, Box } from '@grafana/ui';
 
 import { useGetFrontendSettingsQuery } from '../api';
-import { CONNECT_URL, MIGRATE_URL } from '../constants';
 import { css } from '@emotion/css';
-import { getConfigurationStatus } from './utils';
+import { getConfigurationStatus } from './features';
 import { useState } from 'react';
 import { SetupModal } from './SetupModal';
-import { feature_ini, ngrok_example, root_url_ini } from './types';
+import { FeaturesList } from './FeaturesList';
+import { EnhancedFeatures } from './EnhancedFeatures';
 
 type SetupType = 'public-access' | 'required-features' | null;
 
-interface FeatureItemProps {
-  children: React.ReactNode;
-}
+// Configuration examples
+const feature_ini = `# In your custom.ini file
+app_mode = development
 
-const FeatureItem = ({ children }: FeatureItemProps) => (
-  <Text variant="body">
-    <Icon name="check" className="text-success" /> {children}
-  </Text>
-);
+[feature_toggles]
+provisioning = true
+kubernetesDashboards = true
+unifiedStorageSearch = true
+kubernetesClientDashboardsFolders = true
 
-interface FeatureCardProps {
-  title: string;
-  description: string;
-  icon?: React.ReactNode;
-  action?: React.ReactNode;
-  showBorder?: boolean;
-}
+# If you want easy kubectl setup development mode
+grafanaAPIServerEnsureKubectlAccess = true`;
 
-interface IconCircleProps {
-  icon: string;
-  color: string;
-  background: string;
-}
+const ngrok_example = `ngrok http 3000
 
-const IconCircle = ({ icon, color, background }: IconCircleProps) => (
-  <div
-    className={css`
-      background: ${background};
-      border-radius: 50%;
-      padding: 16px;
-      width: fit-content;
-    `}
-  >
-    <Icon name={icon as IconName} size="xxl" color={color} />
-  </div>
-);
+Help shape K8s Bindings https://ngrok.com/new-features-update?ref=k8s
 
-const FeatureCard = ({ title, description, icon, action, showBorder = false }: FeatureCardProps) => (
-  <Box width="25%" padding={2}>
-    <div
-      className={css`
-        ${showBorder ? 'border-right: 1px solid rgba(204, 204, 220, 0.15);' : ''}
-        height: 100%;
-      `}
-    >
-      <Stack direction="column" gap={2}>
-        {icon}
-        <Text variant="h3">{title}</Text>
-        <Text variant="body">{description}</Text>
-        {action && <Box>{action}</Box>}
-      </Stack>
-    </div>
-  </Box>
-);
+Session Status                online
+Account                       Roberto Jiménez Sánchez (Plan: Free)
+Version                       3.18.4
+Region                        Europe (eu)
+Latency                       44ms
+Web Interface                 http://127.0.0.1:4040
+Forwarding                    https://d60d-83-33-235-27.ngrok-free.app -> http://localhost:3000
+Connections                   ttl     opn     rt1     rt5     p50     p90
+                              50      2       0.00    0.00    83.03   90.56
 
-interface FeaturesListProps {
-  hasPublicAccess: boolean;
-  hasImageRenderer: boolean;
-  hasRequiredFeatures: boolean;
-  onMigrate: () => void;
-  onConnect: () => void;
-  onSetupFeatures: () => void;
-}
+HTTP Requests
+-------------
 
-const FeaturesList = ({
-  hasPublicAccess,
-  hasImageRenderer,
-  hasRequiredFeatures,
-  onMigrate,
-  onConnect,
-  onSetupFeatures,
-}: FeaturesListProps) => (
-  <Stack direction="column" gap={2}>
-    <Text variant="h2">Provisioning as-code directly from Grafana</Text>
-    <FeatureItem>
-      Manage your dashboards as code and deploy them automatically from your GitHub repository or local storage
-    </FeatureItem>
-    <FeatureItem>
-      Review, discuss, and approve dashboard changes with your team before they go live using GitHub pull requests
-    </FeatureItem>
-    <FeatureItem>
-      Export your existing dashboards as code and store them in GitHub repositories for version control and
-      collaboration
-    </FeatureItem>
-    {hasPublicAccess && (
-      <FeatureItem>
-        Automatically provision and update your dashboards as soon as changes are pushed to your GitHub repository
-      </FeatureItem>
-    )}
-    {hasImageRenderer && hasPublicAccess && (
-      <FeatureItem>Visual previews in pull requests to review your changes before going live</FeatureItem>
-    )}
+09:18:46.147 CET             GET  /favicon.ico                   302 Found
+09:18:46.402 CET             GET  /login`;
 
-    <LinkButton fill="text" href="#" icon="external-link-alt">
-      Learn more
-    </LinkButton>
-
-    {hasRequiredFeatures ? (
-      <>
-        <Stack direction="row" alignItems="center" gap={2}>
-          <Button size="md" icon="plus" onClick={onMigrate}>
-            Migrate Grafana to repository
-          </Button>
-          <Text variant="body">or</Text>
-          <LinkButton fill="outline" icon="plus" onClick={onConnect}>
-            Connect Grafana to repository
-          </LinkButton>
-        </Stack>
-      </>
-    ) : (
-      <Box>
-        <LinkButton fill="outline" onClick={onSetupFeatures}>
-          Set up required features
-        </LinkButton>
-      </Box>
-    )}
-  </Stack>
-);
-
-interface EnhancedFeaturesProps {
-  hasPublicAccess: boolean;
-  hasImageRenderer: boolean;
-  onSetupPublicAccess: () => void;
-}
-
-const EnhancedFeatures = ({ hasPublicAccess, hasImageRenderer, onSetupPublicAccess }: EnhancedFeaturesProps) => (
-  <Box marginTop={2}>
-    <Text variant="h2">Unlock enhanced functionality for GitHub</Text>
-    <Box marginTop={4}>
-      <Stack direction="row" gap={2}>
-        <FeatureCard
-          title="Instantenous provisioning"
-          description="Automatically provision and update your dashboards as soon as changes are pushed to your GitHub repository"
-          icon={<IconCircle icon="sync" color="primary" background="rgba(24, 121, 219, 0.12)" />}
-          action={
-            !hasPublicAccess && (
-              <LinkButton fill="outline" onClick={onSetupPublicAccess}>
-                Set up public access
-              </LinkButton>
-            )
-          }
-          showBorder
-        />
-        <FeatureCard
-          title="Visual previews in pull requests"
-          description="Review how your changes look like before going live in Grafana and directly in pull requests"
-          icon={
-            <Stack direction="row" gap={2}>
-              <IconCircle icon="camera" color="orange" background="rgba(255, 120, 10, 0.12)" />
-              <IconCircle icon="code-branch" color="purple" background="rgba(135, 73, 237, 0.12)" />
-            </Stack>
-          }
-          action={
-            !hasImageRenderer && (
-              <LinkButton
-                fill="outline"
-                href="https://grafana.com/grafana/plugins/grafana-image-renderer/"
-                icon="external-link-alt"
-              >
-                Set up image rendering
-              </LinkButton>
-            )
-          }
-        />
-      </Stack>
-    </Box>
-  </Box>
-);
+const root_url_ini = `[server]
+root_url = https://d60d-83-33-235-27.ngrok-free.app`;
 
 export default function GettingStarted() {
   const settingsQuery = useGetFrontendSettingsQuery();
   const legacyStorage = settingsQuery.data?.legacyStorage;
-  const navigate = useNavigate();
 
   const { hasPublicAccess, hasImageRenderer, hasRequiredFeatures } = getConfigurationStatus();
   const [showInstructionsModal, setShowModal] = useState(false);
@@ -250,14 +114,6 @@ export default function GettingStarted() {
             hasPublicAccess={hasPublicAccess}
             hasImageRenderer={hasImageRenderer}
             hasRequiredFeatures={hasRequiredFeatures}
-            onMigrate={async () => {
-              await settingsQuery.refetch();
-              navigate(MIGRATE_URL);
-            }}
-            onConnect={async () => {
-              await settingsQuery.refetch();
-              navigate(CONNECT_URL);
-            }}
             onSetupFeatures={() => {
               setSetupType('required-features');
               setShowModal(true);
